requirements.txt
streamlit>=1.28.0
pydantic>=2.0.0
python-dateutil>=2.8.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=0.1.0
langsmith>=0.1.0
google-generativeai>=0.7.0
openai>=1.30.0
pymongo>=4.6.0
